+++ NEWS.md
@@ -16,8 +16,12 @@ copyright (c) 2001-2020, Karl W Broman
     at https://www.r-project.org/Licenses/GPL-3
 ----------------------------------------------------------------------
 
-## Version 1.47, 2020-05-03
+## Version 1.47, 2020-05-13
 
+### Minor changes
+
+- Added function `find_large_intervals()` for finding inter-marker
+  intervals in a map with length greater than some value.
 
 
 ## Version 1.46, 2020-02-28